+++ src/Nakama/ISocketAdapter.cs
@@ -1,28 +1,27 @@
-/**
- * Copyright 2019 The Nakama Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
+// Copyright 2019 The Nakama Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 
 using System;
 using System.Threading;
+using System.Threading.Tasks;
 
 namespace Nakama
 {
     /// <summary>
     /// An adapter which implements a socket with a protocol supported by Nakama.
     /// </summary>
-    public interface ISocketAdapter : IDisposable
+    public interface ISocketAdapter
     {
         /// <summary>
         /// An event dispatched when the socket is connected.
@@ -57,21 +56,21 @@ namespace Nakama
         /// <summary>
         /// Close the socket with an asynchronous operation.
         /// </summary>
-        void Close();
+        Task CloseAsync();
 
         /// <summary>
         /// Connect to the server with an asynchronous operation.
         /// </summary>
         /// <param name="uri">The URI of the server.</param>
         /// <param name="timeout">The timeout for the connect attempt on the socket.</param>
-        void Connect(Uri uri, int timeout);
+        Task ConnectAsync(Uri uri, int timeout);
 
         /// <summary>
         /// Send data to the server with an asynchronous operation.
         /// </summary>
         /// <param name="buffer">The buffer with the message to send.</param>
-        /// <param name="cancellationToken">A cancellation token used to propagate when the operation should be canceled.</param>
         /// <param name="reliable">If the message should be sent reliably (will be ignored by some protocols).</param>
-        void Send(ArraySegment<byte> buffer, CancellationToken cancellationToken, bool reliable = true);
+        /// <param name="canceller">A cancellation token used to propagate when the operation should be canceled.</param>
+        Task SendAsync(ArraySegment<byte> buffer, bool reliable = true, CancellationToken canceller = default);
     }
 }