+++ tests/Nakama.Tests/AwaitedSocketTaskTest.cs
@@ -1,20 +1,19 @@
-/**
- * Copyright 2019 The Nakama Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
+// Copyright 2019 The Nakama Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 
 using System;
+using System.Net.Sockets;
 using System.Threading.Tasks;
 using Xunit;
 
@@ -31,11 +30,7 @@ namespace Nakama.Tests
             _socket = Nakama.Socket.From(_client);
         }
 
-        public void Dispose()
-        {
-            _socket.Dispose();
-            _client = null;
-        }
+        public void Dispose() => _client = null;
 
         [Fact(Timeout = TestsUtil.TIMEOUT_MILLISECONDS)]
         public async void Socket_AwaitedTasks_AreCanceled()
@@ -52,7 +47,7 @@ namespace Nakama.Tests
         }
 
         [Fact(Timeout = TestsUtil.TIMEOUT_MILLISECONDS)]
-        public async void Socket_AwaitedTasksAfterDisconnect_AreCanceled()
+        public async void Socket_AwaitedTasksAfterDisconnect_ThrowException()
         {
             var id = Guid.NewGuid().ToString();
             var session = await _client.AuthenticateCustomAsync(id);
@@ -62,7 +57,7 @@ namespace Nakama.Tests
             var statusTask1 = _socket.FollowUsersAsync(new[] {session.UserId});
             var statusTask2 = _socket.FollowUsersAsync(new[] {session.UserId});
 
-            await Assert.ThrowsAsync<TaskCanceledException>(() => Task.WhenAll(statusTask1, statusTask2));
+            await Assert.ThrowsAsync<SocketException>(() => Task.WhenAll(statusTask1, statusTask2));
         }
     }
 }