+++ tests/Nakama.Tests/Socket/WebSocketTest.cs
@@ -1,37 +1,39 @@
-/**
- * Copyright 2020 The Nakama Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
+// Copyright 2021 The Nakama Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+using System;
+using System.Threading.Tasks;
+using Xunit;
+using Xunit.Abstractions;
 
 namespace Nakama.Tests.Socket
 {
-    using System;
-    using System.Net.Sockets;
-    using System.Threading.Tasks;
-    using Xunit;
-
     public class WebSocketTest
     {
-        private IClient _client;
-        private ISocket _socket;
+        private readonly ITestOutputHelper _testOutputHelper;
+        private readonly IClient _client;
+        private readonly ISocket _socket;
 
         // ReSharper disable RedundantArgumentDefaultValue
 
-        public WebSocketTest()
+        public WebSocketTest(ITestOutputHelper testOutputHelper)
         {
+            _testOutputHelper = testOutputHelper;
             _client = TestsUtil.FromSettingsFile();
             _socket = Nakama.Socket.From(_client);
+            var logger = new StdoutLogger();
+            _socket.ReceivedError += e => logger.ErrorFormat(e.Message);
         }
 
         [Fact(Timeout = TestsUtil.TIMEOUT_MILLISECONDS)]
@@ -81,12 +83,12 @@ namespace Nakama.Tests.Socket
         }
 
         [Fact(Timeout = TestsUtil.TIMEOUT_MILLISECONDS)]
-        public async Task MultipleConnectAttemptsThrowException()
+        public async Task MultipleConnectAttemptsDoesNotThrowException()
         {
             var session = await _client.AuthenticateCustomAsync($"{Guid.NewGuid()}");
             await _socket.ConnectAsync(session);
             Assert.True(_socket.IsConnected);
-            await Assert.ThrowsAsync<SocketException>(() => _socket.ConnectAsync(session));
+            await _socket.ConnectAsync(session);
         }
 
         [Fact(Timeout = TestsUtil.TIMEOUT_MILLISECONDS)]
@@ -96,6 +98,16 @@ namespace Nakama.Tests.Socket
             await _socket.CloseAsync();
             await _socket.ConnectAsync(session);
             Assert.True(_socket.IsConnected);
+        }
+
+        [Fact(Skip = "Test case requires 60 seconds minimum execution time.")]
+        public async Task LongLivedSocketLifecycle()
+        {
+            var session = await _client.AuthenticateCustomAsync($"{Guid.NewGuid()}");
+            await _socket.ConnectAsync(session, false, 5);
+            await Task.Delay(TimeSpan.FromSeconds(60));
+            Assert.True(_socket.IsConnected);
+            _ = _socket.CloseAsync();
         }
     }
 }